Move ReadExt to decode module
Reading is part of decoding, the `ReadExt` should live in the new decode
module, move it there.

Add public re-export to `encode` and do not change usage of the
extension trait anywhere else in the codebase.
